+++ CHANGELOG.md
@@ -72,6 +72,10 @@ and this project adheres to
 
 ## [15.1.1] – 2026-08-05
 
+### Fixed
+
+- GDKPixbuf JPEG output is no longer empty with GDKPixbuf 2.44.
+
 ### Changed
 
 - When using `dl_iterate_phdr` to discover the plugin directory, the default